Foam breath, spawning, low health, healing Added foam breath prototype ability.
Adds a hitbox for the ability to the environment which will slow enemies
and player. Does not auto-despawn yet.

Spawn points were made to spawn enemies at intervals, and not if there
is anything in the spawn location (hitbox)

A low health indicator (red circle on ground) was added to display when
the user has low health.

Functionality was added to player health script to facilitate health
regeneration after a specified amount of time, and at a specified rate.
Being damaged stops the regeneration.